Extract build-config tests to new file
